backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-backend"
version = "0.1.0"
description = "Project Atlas backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.8",
  "uvicorn[standard]>=0.34.0",
  "pydantic-settings>=2.7.1",
  "sqlalchemy>=2.0.38",
  "psycopg[binary]>=3.2.5",
  "bcrypt>=4.2.1",
  "python-jose[cryptography]>=3.3.0",
  "alembic>=1.14.1",
  "httpx>=0.28.1",
  "python-multipart>=0.0.20",
  "Pillow>=11.1.0"
]

[dependency-groups]
dev = [
  "ruff>=0.9.7",
  "pytest>=8.3.4",
  "httpx>=0.28.1"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.setuptools.packages.find]
include = ["app*"]
